+++ src/main/java/com/main/student/inventory/studentInventory/repository/ReceiptInfoRepository.java
@@ -3,5 +3,5 @@ package com.main.student.inventory.studentInventory.repository;
 import com.main.student.inventory.studentInventory.model.ReceiptInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceiptInfoRepository extends JpaRepository<ReceiptInfo, Long> {
+public interface ReceiptInfoRepository extends JpaRepository<ReceiptInfo, Integer> {
 }
